+++ R/replace_special_characters_internal.R
@@ -5,7 +5,12 @@
 #' need to be elements of \code{stringi::stri_trans_list()} (like "Latin-ASCII", which is often useful) or names of lookup tables (currently
 #' only "german" is supported). In the order of the entries the letters of the input
 #'  string will be transliterated via \code{stringi::stri_trans_general()} or replaced via the 
-#'  matches of the lookup table.
+#'  matches of the lookup table. When named character elements are supplied as part of `transliterations`, anyting that mateches the names is replaced by the corresponding value.
+
+#' You should use this feature with care in case of \code{case = "parsed"}, \code{case = "internal_parsing"} and 
+#' \code{case = "none"}, since for upper case letters, which have transliterations/replacements
+#'  of length 2, the second letter will be transliterated to lowercase, for example Oe, Ae, Ss, which
+#'  might not always be what is intended. In this case you can make usage of the option to supply named elements and specify the transliterations yourself.
 #' 
 #' @param case Length one character, from the input options of \code{to_any_case}.
 #' 
@@ -37,13 +42,15 @@ replace_special_characters_internal <- function(string, transliterations, case){
                 "\u00E5" = "a")
     )
   for (i in seq_along(transliterations)){
-    if(transliterations[i] %in% stringi::stri_trans_list()){
-      string <- stringi::stri_trans_general(string, transliterations[i])
-      } else if(transliterations[i] %in% names(dictionary)){
+    if(isTRUE(!is.null(names(transliterations)[i]) &
+              names(transliterations)[i] != "")){
+      string <- stringr::str_replace_all(string, transliterations[i])
+    } else if(transliterations[i] %in% names(dictionary)){
         string <- stringr::str_replace_all(string, dictionary[[transliterations[i]]]) 
+      } else if(transliterations[i] %in% stringi::stri_trans_list()){
+        string <- stringi::stri_trans_general(string, transliterations[i])
       } else {
-          stop("Input to `transliterations` must be `NULL` or a string containing elements
-               from the internal lookup dictionaries or from `stringi::stri_trans_list()`.",
+          stop("Input to `transliterations` must be `NULL`, a string containing elements from the internal lookup dictionaries or from `stringi::stri_trans_list()` or a named vector.",
                   call. = FALSE)
       }
   }