+++ NEWS.md
@@ -1,7 +1,8 @@
 # snakecase 0.9.2
 
-* added `to_sentence_case()` (same as snake, but first letter is uppercase and default sep_out is space).
-* added `numerals` argument to all caseconverters including `to_any_case()` to format the alignment of digits (`middle`, `left`,`right`, `asis`). Therefore `parsing_option` nr 4 might be removed in later releases, as it is the same as `parsing_option = 1` and `numerals = "asis"`.
+* **cases**: added `to_sentence_case()` (same as snake, but first letter is uppercase and default sep_out is space).
+* **numerals**: added `numerals` argument to all caseconverters including `to_any_case()` to format the alignment of digits (`middle`, `left`,`right`, `asis`). Therefore `parsing_option` nr 4 might be removed in later releases, as it is the same as `parsing_option = 1` and `numerals = "asis"`.
+* **transliterations**: When named character elements are supplied as part of `transliterations`, anyting that mateches the name is replaced by the corresponding value.
 
 # snakecase 0.9.1